+++ frontend/projects/api/src/lib/deployment/version.service.interface.ts
@@ -1,0 +1,5 @@
+import { DeploymentVersion } from '../models/deployment-version';
+
+export interface IVersionService {
+  get(): Promise<DeploymentVersion>;
+}